% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_ghqc_create_app}
\alias{run_ghqc_create_app}
\title{Initialize QC checklists for a set of files on Github}
\usage{
run_ghqc_create_app(
  app_name = "ghqc_create_app",
  qc_dir = getwd(),
  lib_path = "~/ghqc-rpkgs/"
)
}
\arguments{
\item{app_name}{the name of the app to run in the background}

\item{qc_dir}{the directory in which the app is run}

\item{lib_path}{the path to the ghqc package and its dependencies}
}
\description{
This function allows the user to initialize QC checklists for a set of files on Github.

To initialize QC checklists, the user:
\enumerate{
\item names a new QC milestone or selects an existing on to coalesce a set of files to be QCed,
\item optionally sets a milestone description if the user opts to create a new milestone,
\item optionally selects a set of collaborators to assign to various checklists in the QC milestone,
\item selects a set of files from the file tree on the right side of the pane,
\item allocates a checklist type to each file, along with an optional assignee, and finally
\item posts the QC milestone by clicking "Create QC Items" on the bottom of the pane.
}

At any time, the user can:
\itemize{
\item Click \verb{checklist info} to the right of the file tree to view the items in each checklist
\item Click \code{preview} next to any file to view its current contents
}
}
\examples{
# example code
run_ghqc_create_app()

}
\seealso{
\link{run_ghqc_update_app} and \link{run_ghqc_report_app}
}
