% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_ghqc_report_app}
\alias{run_ghqc_report_app}
\title{Generate a report for a set of QC milestones}
\usage{
run_ghqc_report_app(
  app_name = "ghqc_report_app",
  qc_dir = getwd(),
  lib_path = "~/ghqc-rpkgs/"
)
}
\arguments{
\item{app_name}{the name of the app to run in the background}

\item{qc_dir}{the directory in which the app is run}

\item{lib_path}{the path to the ghqc package and its dependencies}
}
\description{
This function allows the user to generate a report for a set of QC milestones on Github,
created with the \link{run_ghqc_create_app}, whose issues have been optionally updated with the \link{run_ghqc_update_app}.

To generate a QC report, the user selects the set of QC milestones to include in the report.

The user can optionally:
\enumerate{
\item give a name for the pdf, else the report will be named after the remote repository and milestones
\item give the directory in which to generate the pdf
\item indicate whether the report should only include the summary table of QC issues for each milestone
}

Finally, the user clicks "Generate QC Report" at the bottom of the pane.
}
\examples{
# example code
run_ghqc_report_app()

}
\seealso{
\link{run_ghqc_create_app} and \link{run_ghqc_update_app}
}
