% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_ghqc_update_app}
\alias{run_ghqc_update_app}
\title{Update a QC issue with script changes amid QC}
\usage{
run_ghqc_update_app(
  app_name = "ghqc_update_app",
  qc_dir = getwd(),
  lib_path = "~/ghqc-rpkgs/"
)
}
\arguments{
\item{app_name}{the name of the app to run in the background}

\item{qc_dir}{the directory in which the app is run}

\item{lib_path}{the path to the ghqc package and its dependencies}
}
\description{
This function allows a user to update a QC issue with a comment that gives context surrounding changes to the QC file.
An update comment indicates the hash and git sha of the updated file and an optionally displays the file difference between versions.

To use this app, first initialize a set of QC issues with \link{run_ghqc_create_app}.

To update a QC issue, the user:
\enumerate{
\item optionally selects the milestone of the relevant QC issue (for ease of filtering)
\item selects the QC issue they'd like to update
\item optionally gives a contextualizing message about the file changes
\item indicates whether or not display the file difference
\item chooses whether to
\itemize{
\item compare the original version with the most recent or,
\item manually select an older file version (reference) to compare with a newer file version (comparator)
}
\item optionally previews the comment before posting it, and finally
\item posts the comment on the Github QC issue
}
}
\examples{
# example code
run_ghqc_update_app()

}
\seealso{
\link{run_ghqc_create_app} and \link{run_ghqc_report_app}
}
